IPython Console: Rename widgets/jupyter_qtconsole to widgets/ipythonconsole
